add duration
