+++ scripts/browse-card/browse-card.js
@@ -503,119 +503,22 @@ const getVideoClipModal = () => {
   return videoClipModalPromise;
 };
 
-/**
- * Decorates upcoming event cards with additional features
- * @param {HTMLElement} card - The card element to decorate
- * @param {Object} model - The data model for the card
- */
-const decorateUpcomingEvents = (card, model) => {
-  if (!card || !model || model.contentType?.toLowerCase() !== CONTENT_TYPES.UPCOMING_EVENT.MAPPING_KEY) return;
+// Dynamic imports for event decorators
+let upcomingEventsPromise = null;
+let onDemandEventsPromise = null;
 
-  loadCSS(`${window.hlx.codeBasePath}/scripts/browse-card/browse-card-upcoming-events.css`);
-
-  // Remove the browse-card-options div from upcoming event cards
-  const cardOptions = card.querySelector('.browse-card-options');
-  if (cardOptions) {
-    cardOptions.remove();
+const getUpcomingEventsDecorator = () => {
+  if (!upcomingEventsPromise) {
+    upcomingEventsPromise = import('./browse-card-upcoming-events.js');
   }
-
-  const cardFigure = card?.querySelector('.browse-card-figure');
-  if (!cardFigure) return;
-
-  const { event } = model;
-  const hasSpeakers = event?.speakers?.name?.length > 0 && event?.speakers?.profilePictureURL?.length > 0;
-  const hasSeries = event?.series;
-  const seriesText = hasSeries ? event?.series : null;
-
-  if (hasSpeakers) {
-    let speakersContainer = cardFigure?.querySelector('.event-speakers-container');
-    if (!speakersContainer) {
-      speakersContainer = createTag('div', { class: 'event-speakers-container' });
-      cardFigure.appendChild(speakersContainer);
-    } else {
-      speakersContainer.innerHTML = '';
-    }
-
-    const speakerNames = Array.isArray(event?.speakers?.name)
-      ? event?.speakers?.name
-      : event?.speakers?.name?.split(',')?.map((name) => name?.trim());
-
-    const profilePictures = Array.isArray(event?.speakers?.profilePictureURL)
-      ? event?.speakers?.profilePictureURL
-      : event?.speakers?.profilePictureURL?.split(',')?.map((url) => url?.trim());
-
-    const maxSpeakers = Math.min(speakerNames?.length || 0, profilePictures?.length || 0, 3);
-    for (let i = 0; i < maxSpeakers; i += 1) {
-      const speakerImgContainer = createTag('div', { class: 'speaker-profile-container' });
-      const speakerImg = createTag('img', {
-        src: profilePictures?.[i],
-        alt: speakerNames?.[i],
-        class: 'speaker-profile-image',
-      });
-      speakerImgContainer.appendChild(speakerImg);
-      speakersContainer.appendChild(speakerImgContainer);
-    }
-  } else if (hasSeries) {
-    const existingSpeakers = cardFigure?.querySelector('.event-speakers-container');
-    if (existingSpeakers) {
-      existingSpeakers.remove();
-    }
-  }
-
-  if (seriesText) {
-    const existingSeriesBanner = cardFigure?.querySelector('.event-series-banner, .event-series-banner-no-speakers');
-    if (existingSeriesBanner) {
-      existingSeriesBanner.remove();
-    }
-
-    const bannerClass = hasSpeakers ? 'event-series-banner' : 'event-series-banner event-series-banner-no-speakers';
-
-    if (!cardFigure?.querySelector(`.${bannerClass.replace(' ', '.')}`)) {
-      cardFigure.appendChild(createTag('div', { class: bannerClass }, seriesText));
-    }
-  }
-
-  if (event?.type) {
-    const locationType = event?.type;
-    if (locationType && !cardFigure?.querySelector('.location-type')) {
-      cardFigure.appendChild(createTag('div', { class: 'location-type' }, locationType));
-    }
-  }
+  return upcomingEventsPromise;
 };
 
-const decorateOnDemandEvents = (card, model) => {
-  if (!card || !model || model.contentType?.toLowerCase() !== CONTENT_TYPES.EVENT.MAPPING_KEY) return;
-
-  loadCSS(`${window.hlx.codeBasePath}/scripts/browse-card/browse-card-on-demand-events.css`);
-
-  const { event } = model;
-  const cardFigure = card.querySelector('.browse-card-figure');
-  if (!cardFigure) return;
-
-  const cardContent = card.querySelector('.browse-card-content');
-
-  cardFigure.querySelector('.laptop-container')?.remove();
-  const img = cardFigure.querySelector('img');
-  img?.remove();
-
-  cardFigure.querySelector('.play-button')?.remove();
-
-  img?.addEventListener('load', () => {
-    cardFigure.querySelector('.play-button')?.remove();
-  });
-
-  cardFigure.querySelector('.event-series-banner')?.remove();
-
-  buildEventContent({
-    event,
-    contentType: model.contentType,
-    cardContent,
-    card,
-  });
-
-  const seriesText = event?.series || 'On Demand Event Series';
-  const banner = createTag('div', { class: 'event-series-banner' }, seriesText);
-  cardFigure.appendChild(banner);
+const getOnDemandEventsDecorator = () => {
+  if (!onDemandEventsPromise) {
+    onDemandEventsPromise = import('./browse-card-on-demand-events.js');
+  }
+  return onDemandEventsPromise;
 };
 
 /**
@@ -908,11 +811,17 @@ export async function buildCard(element, model) {
 
   const cardEl = element.querySelector('.browse-card');
   if (cardEl && (v2 || isBrowseFiltersUpcoming)) {
-    decorateUpcomingEvents(cardEl, model);
+    // Dynamically import and use the upcoming events decorator
+    getUpcomingEventsDecorator().then(({ decorateUpcomingEvents }) => {
+      decorateUpcomingEvents(cardEl, model);
+    });
   }
 
   if (model.contentType?.toLowerCase() === CONTENT_TYPES.EVENT.MAPPING_KEY && isFeatureEnabled('isEventsV2')) {
     const cardElement = element.querySelector('.browse-card');
-    decorateOnDemandEvents(cardElement, model);
+    // Dynamically import and use the on-demand events decorator
+    getOnDemandEventsDecorator().then(({ decorateOnDemandEvents }) => {
+      decorateOnDemandEvents(cardElement, model);
+    });
   }
 }
